[KT, TS] deleted utils
